Bug fix for pressing backspace with empty string

+++ src/fpsGame/Ship.java
@@ -11,7 +11,7 @@ public abstract class Ship extends GameObject{
 	float qsize;
 	float shield;
 	float fireRate;
-	float score;
+	int score;
 
 	Ship(PApplet p, float size, PVector pos){
 		super(p);
